school_admin_bot/requirements.txt
python-telegram-bot[job-queue]==21.0
anthropic>=0.49.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyMuPDF>=1.24.0